# api_service/app/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.BoltMaterialsDTO:
    properties:
      flangeId:
        type: string
      title:
        type: string
    required:
    - flangeId
    - title
    type: object
...
  models.DataResponse:
    properties:
      count:
        type: integer
      data: {}
    type: object
...
  models.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.IdResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
  models.MaterialsDTO:
    properties:
      title:
        type: string
      typeMat:
        type: string
    required:
    - title
    - typeMat
    type: object
...
  models.UpdateMatDTO:
    properties:
      change:
        type: string
      materials:
        items:
          $ref: '#/definitions/proto.AddMaterials'
        type: array
      type:
        type: string
    required:
    - materials
    - type
    type: object
...
  proto.AddMaterials:
    properties:
      short:
        type: string
      title:
        type: string
    type: object
...
  proto.BoltMaterials:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
...
  proto.Materials:
    properties:
      id:
        type: string
      title:
        type: string
      typeMat:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Service for Sealur
  title: Sealur
  version: 0.14.0
paths:
...
  /sealur-pro/additionals/{id}/mat:
    patch:
      consumes:
      - application/json
      description: обновление материалов
      parameters:
      - description: additional materials info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMatDTO'
      - description: addit id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Mat
      tags:
      - Sealur Pro -> additionals
...
  /sealur-pro/bolt-materials:
    get:
      consumes:
      - application/json
      description: получение материалов болтов/шпилек (для опроса)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/proto.BoltMaterials'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Bolt Materials
      tags:
      - Sealur Pro -> bolt materials
    post:
      consumes:
      - application/json
      description: создание материала болтов/шпилек (для опроса)
      parameters:
      - description: bolt materials info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.BoltMaterialsDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Bolt Materials
      tags:
      - Sealur Pro -> bolt materials
  /sealur-pro/bolt-materials/{id}:
    delete:
      consumes:
      - application/json
      description: удаление материала болтов/шпилек (для опроса)
      parameters:
      - description: bolt material id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Bolt Materials
      tags:
      - Sealur Pro -> bolt materials
    put:
      consumes:
      - application/json
      description: обновление материала болтов/шпилек (для опроса)
      parameters:
      - description: bolt material info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.BoltMaterialsDTO'
      - description: bolt material id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Bolt Materials
      tags:
      - Sealur Pro -> bolt materials
...
  /sealur-pro/materials:
    get:
      consumes:
      - application/json
      description: получение материалов (для опроса)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/proto.Materials'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Materials
      tags:
      - Sealur Pro -> materials
    post:
      consumes:
      - application/json
      description: создание материала (для опроса)
      parameters:
      - description: materials info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.MaterialsDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Materials
      tags:
      - Sealur Pro -> materials
  /sealur-pro/materials/{id}:
    delete:
      consumes:
      - application/json
      description: удаление материала (для опроса)
      parameters:
      - description: material id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Materials
      tags:
      - Sealur Pro -> materials
    put:
      consumes:
      - application/json
      description: обновление материала (для опроса)
      parameters:
      - description: material info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.MaterialsDTO'
      - description: material id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Materials
      tags:
      - Sealur Pro -> materials
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"